backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
numpy
matplotlib
python-jose
passlib
requests